Correct issue with removal of Todo Items
Todo Items where not properly remove before, these changes fix the
issue by recording and using new Id into the Todo in the Array List.

+++ app/src/main/java/com/amberved/todo/todoapp/Todo.java
@@ -9,16 +9,16 @@ public class Todo implements Serializable {
 
     private String mText;
     private int mPriority;
-    private int mId;
+    private long mId;
     private String mDueDate;
 
     private static final long serialVersionUID = 5177222050535318633L;
 
-    public int getId() {
+    public long getId() {
         return mId;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.mId = id;
     }
